pyproject.toml
[tool.poetry]
authors = ["Andre Sionek <andre@sionek.com>"]
name = " Chapter: Infrastructure as Code"
description = "Code for the book Real Life Infrastructure as Code with AWS CDK" # noqa
keywords = ['cdk', 'iac', 'infrastructure', 'devops', 'python', 'backend', "platform"] # noqa
readme = "README.md"
homepage = "https://real-life-infrastructure-as-code.com/"
version = "0.0.0"


[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.99.0"


